Large structural change to data and feature extraction
feat_*.m functions are restructured to accept a single data
structure from a segment file as their input.
A new function, computeFeatM.m, performs the previous operation
of looping over all data files for a given subject and ictal type
in a large parfor loop.
ICA handling is now generalised by applying a modelling function to
the raw data after it is loaded (in this case a matrix
multiplication), by computeFeatM.m before it is passed to feat_*.
Functions for adding features to the HDF5 also updated.