ENH: Add support for additional classifiers
Previously several classifiers in sklearn could not be used because
of our use of the sample_weight input to the classifier (which was
never even actually used) which was only available for a subset of
the classifiers.

In more recent versions of sklearn, sample_weight support has been
added to these classifiers, so support for the following classifiers
is added to the settings parsing pipeline:
    LogisticRegression
    RidgeClassifier
    RandomTreesEmbedding
    GradientBoostingClassifier
    SGDClassifier

In sklearn 0.17, sample_weight support was added to
LogisticRegression and RidgeClassifier, so the requirements.txt
has been updated to reflect this.